requirements.txt
# LLM API clients
openai>=1.30.0

# Retrieval toolkit
pyserini>=0.25.0
faiss-cpu>=1.7.4

# Evaluation
pytrec-eval-terrier>=0.5.6

# Data & utilities
datasets>=2.19.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0
jsonlines>=4.0.0

